Fix wrong cast

+++ core/src/datamatrix/DMDetector.cpp
@@ -92,9 +92,9 @@ inline static bool IsValidPoint(const ResultPoint& p, int imgWidth, int imgHeigh
 inline static float RoundToNearest(float x)
 {
 #if defined(__ANDROID__) && defined(__GNUC__)
-	return static_cast<int>(round(x));
+	return static_cast<float>(round(x));
 #else
-	return static_cast<int>(std::round(x));
+	return std::round(x);
 #endif
 }
